Move main src to core
