+++ clean-element/package.json
@@ -20,7 +20,8 @@
   "devDependencies": {
     "@compositor/x0": "^3.1.1",
     "ava": "^0.25.0",
-    "babel-cli": "^6.24.1"
+    "babel-cli": "^6.24.1",
+    "styled-components": "^3.1.6"
   },
   "ava": {
     "require": [
