+++ clean-element/test.js.md
@@ -4,8 +4,14 @@ The actual snapshot is saved in `test.js.snap`.
 
 Generated by [AVA](https://ava.li).
 
-## it removes props
+## does not remove props without propTypes
 
 > Snapshot 1
 
-    '<div class="sc-bdVaJa ibTDPh" data-reactroot=""></div>'
+    '<div class="sc-bwzfXH bvRIsA" id="beep" align="center" data-reactroot="">hello</div>'
+
+## it removes props with prop types
+
+> Snapshot 1
+
+    '<div class="sc-bdVaJa gooaBC" id="beep" data-reactroot="">hello</div>'
