+++ core/.npmignore
@@ -1,0 +1,9 @@
+docs
+src
+.nyc_output
+coverage
+coverage.lcov
+bench
+.travis.yml
+test.js
+webpack.config.js
