+++ core/package.json
@@ -1,0 +1,75 @@
+{
+  "name": "styled-system",
+  "version": "1.1.4",
+  "description": "Design system utilities for styled-components, glamorous, and other css-in-js libraries",
+  "main": "dist/index.js",
+  "scripts": {
+    "prepublish": "babel src -d dist",
+    "start": "x0 dev docs/App.js -op 8888",
+    "build": "x0 build docs/App.js -d docs --static --cssLibrary='styled-components'",
+    "bench": "node bench",
+    "size": "bundlesize",
+    "cover": "nyc report --reporter=html --reporter=lcov > coverage.lcov && codecov",
+    "test": "nyc ava"
+  },
+  "keywords": [
+    "react",
+    "css-in-js",
+    "styled-components",
+    "glamorous"
+  ],
+  "author": "Brent Jackson",
+  "license": "MIT",
+  "devDependencies": {
+    "@compositor/x0": "^3.1.1",
+    "ava": "^0.25.0",
+    "babel-cli": "^6.24.1",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-env": "^1.5.2",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "babel-register": "^6.24.1",
+    "benchmark": "^2.1.4",
+    "bundlesize": "^0.16.0",
+    "codecov": "^2.2.0",
+    "grid-styled": "^2.0.0-2",
+    "microtime": "^2.1.6",
+    "nyc": "^11.0.3",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "react-live": "^1.6.1",
+    "react-test-renderer": "^16.2.0",
+    "remark": "^9.0.0",
+    "remark-react": "^4.0.1",
+    "remark-slug": "^5.0.0",
+    "styled-components": "^3.1.6"
+  },
+  "ava": {
+    "require": [
+      "babel-register"
+    ],
+    "babel": "inherit"
+  },
+  "nyc": {
+    "ignore": [
+      "test.js",
+      "docs",
+      "bench"
+    ]
+  },
+  "bundlesize": [
+    {
+      "path": "dist/*.js"
+    }
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/jxnblk/styled-system.git"
+  },
+  "dependencies": {
+    "prop-types": "^15.6.0"
+  },
+  "peerDependencies": {
+    "react": "^0.14.x || ^15.x || ^16.x"
+  }
+}
