+++ core/src/border-width.js
