+++ core/src/remove-props.js
