+++ core/src/width.js
